+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 29/06/2020
 
-- Change nodejs version requirement to v12.17.
+- Change nodejs version requirement to v12.16.
 
 # v1.5.1